[Platinum IV] Title: 휴대폰 자판, Time: 452 ms, Memory: 124412 KB -BaekjoonHub

+++ 백준/Platinum/5670. 휴대폰 자판/휴대폰 자판.cc
@@ -59,7 +59,7 @@ int main()
         for(int i=0; i<N; i++){
             sum+=trie->solve(v[i]);
         }
-        cout<<sum/double(N)<<"\n";
+        cout<<sum/N<<"\n";
         
         delete trie;
     }
